Rewrite uninstall script in Python. Better now.
The rewritten uninstall script will be smarter
about searching for the binary and the man pages
and it'll first present the paths it found to the
user and ask if they should be removed. It also
sends them to the trash (via Finder) instead of
rm'ing them. Edit some comments in the install
script as well.