Nicer CSS for examples.
Make the examples page look nicer by adjusting the
CSS a bit. Ignore the config file when running
icalBuddy in order to generate the example output
by adding a '-cf ""' argument to the command
(instead of temporarily renaming the config file).
This way I can later add a special config file to
be used just for the example generation.