Rewrite install script in Python.
The new install script should be more intelligent
about finding already installed paths, it'll
expand tildes in user-specified prefixes and
let the user specify a prefix both via the
--prefix=/path argument and within the installer
itself. It will also use Python's built-in shutil
operations for creating folders and copying files
if root access rights are not required and only
shell out to sudo if the former fails.