More correct handling of user-input dates.
Explicitly set the times for user-input dates that
don't include a time specification to 00:00:00 (if
the date specifies the start of a range) or
23:59:59 (if it specifies the end). For date
inputs interpreted as natural language we have
to use very crappy heuristics to figure out
whether the input has included a time
specification, but that's not the end of the
world. Also limit the considered date range better
for recurring events.